Fix newline insertion
This fixes an issue where newlines added to the end of the text would
automatically get trimmed due to d1f6a88 trimming trailing newlines from
the storage file on reload.

To avoid this, we simply add a newline whenever writing to the file,
which both resolves this issue and makes the underlying text file more
consistent with what other editors would commonly produce.