Fix caret color visibility
In 155eaad the caret's color was changed to use the highlight color, to
avoid blending into the previous line's text, however this causes it to
blend inside the previous line's selection with multiline selections.

To ensure the caret is ALWAYS visible, it is now drawn in two colors: An
outline in the highlight color, with a center using the background
color. As a result it should now always be at least partially visible.